docs(core): fix broken demo
